requirements.txt
Flask>=2.0
pandas
numpy
scikit-learn
spacy
pdfminer.six
python-docx
nltk
reportlab
python-dotenv
gunicorn
After installing, run: python -m spacy download en_core_web_sm